requirements.txt
openai-whisper
sentence-transformers
faiss-cpu
numpy
pandas
python-dotenv 